Refactor the C# MockUploader out of PingTypeTest.cs
This will be helpful for Events tests as well.
